Refatoracao das classes estabelecimento e lista, correcoes de bugs nos adapters

+++ app/src/main/java/mobile/listaacessivel/fafica/listaacessvel/adapters/MyArrayAdapterCriarListaPasso3.java
@@ -91,7 +91,7 @@ public class MyArrayAdapterCriarListaPasso3 extends BaseAdapter{
         if(produto != null){
             holder.nome_produto.setText(produto.getDescricao());
             holder.marca_produto.setText("Marca: " + produto.getMarca());
-            holder.valor_produto.setText("Valor: R$ " + Double.toString(produto.getValor()));
+            holder.valor_produto.setText("Valor: R$ " + Float.toString(produto.getValor()));
             holder.selecao.setText(produto.getSelecao());
         }
